requirement.txt
Requirements:

pandas==1.1.5
numpy==1.18.5
pycaret==2.2.0
streamlit==0.71.0
Pillow==8.0.1
